+++ apps/router-e2e/__e2e__/native-tabs/app/_layout.tsx
@@ -19,8 +19,17 @@ export default function Layout() {
           <NativeTabs.Trigger.Label>Index label</NativeTabs.Trigger.Label>
           <NativeTabs.Trigger.Icon sf="applewatch.side.right" md="watch" />
         </NativeTabs.Trigger>
-        <NativeTabs.Trigger name="nested" rippleColor="#0f0" indicatorColor="#ff0">
-          <NativeTabs.Trigger.Icon drawable="ic_lock_open" />
+        <NativeTabs.Trigger
+          name="nested"
+          rippleColor="#0f0"
+          indicatorColor="#ff0"
+          disabled
+          listeners={{
+            tabPress: () => {
+              console.log('disabled tab press');
+            },
+          }}>
+          <NativeTabs.Trigger.Icon drawable="ic_lock_open" sf="lock" />
           <NativeTabs.Trigger.Label hidden={Platform.OS !== 'web'} />
           <NativeTabs.Trigger.Badge>1</NativeTabs.Trigger.Badge>
         </NativeTabs.Trigger>